usando archivo de barril y agregando comentarios descriptivos

+++ src/lib/firebasePublicaciones.js
@@ -1,55 +1,35 @@
-// import { initializeApp } from 'https://www.gstatic.com/firebasejs/9.17.1/firebase-app.js';
-// importo la configuracion de firebase
-import {
-  getFirestore,
-  addDoc,
-  getDocs,
-  doc,
-  collection,
-  deleteDoc
-} from 'https://www.gstatic.com/firebasejs/9.17.1/firebase-firestore.js';
-import FirebaseApp from '../firebaseConfig.js';
-import { getAuth } from 'https://www.gstatic.com/firebasejs/9.17.1/firebase-auth.js';
-
+// importo la configuracion de firebase del archivo de barril
+import Firebase from '../firebaseConfig.js';
+// pongo los nombres usuales de los objetos y funciones de firebase
+const { db, auth, addDoc, getDocs, doc, collection, deleteDoc } = Firebase;
 export const firebaseCrearPublicacion = async (texto) => {
-  // configurando la aplicacion segun datos de la consola de firebase
-  // const app = initializeApp(firebaseConfig,'mifirestore');
-
-  // conectando a la base de datos de firestore
-  const db = getFirestore(FirebaseApp);
+  // insertando la publicacion en la coleccion Publicaciones con el documento publicacion
   await addDoc(collection(db, 'Publicaciones'), { publicacion: texto });
   console.log('dato insertado');
 };
 
 export const firebaseLeerPublicacion = async () => {
-  // configurando la aplicacion segun datos de la consola de firebase
-  // añadi otra instancia de firebase mifirestore que no cause conflicto con el de autentication
-  // const app = initializeApp(firebaseConfig,'mifirestore');
-  // conectando a la base de datos de firestore
-  const db = getFirestore(FirebaseApp);
-  const auth = getAuth();
   // con el await decimos que esperemos que termine la funcion getDocs antes de continuar
+  // leemos todos los documentos de la coleccion Publicaciones
   const querySnapshot = await getDocs(collection(db, 'Publicaciones'));
 
-  // no funciono con for of normal para consultar
-  /* for(const doc of querySnapshot){
-      console.log(doc.data())
-    } */
+  // iniciamos el template String
   let HtmlString = '';
-  for(let i=0;i<querySnapshot.docs.length;i++){
-    const document=querySnapshot.docs[i]
-    console.log(document.id)
-    console.log(document.data());
-    console.log(document.data().publicacion);
-    //leyendo los likes de la publicacion
+  // recorremos todos los documentos de las publicaciones
+  for(let i=0; i<querySnapshot.docs.length; i++){
+    // guardamos cada publicacion en document
+    const document=querySnapshot.docs[i];
+    // seleccionamos la sub coleccion likes
     const likesRef = collection(doc(db, "Publicaciones", document.id), "likes");
+    // leemos los likes de la publicacion
     const likesDePublicacion = await getDocs(likesRef);
     let tieneLike=false
+    // busco si estoy entre los usuarios que dieron like a la publicacion
+    // si estoy cambio el valor de tieneLike a true
     likesDePublicacion.forEach((documentLike) => {
-      // console.log(doc.id, " => ", doc.data().usuario);
       if(documentLike.data().email===auth.currentUser.email) tieneLike=true
     });
-
+    // si di like se pintara el img con un like pintado sino estara vacio segun la variable tieneLike
     HtmlString += `
       <article class='miPublicacion'>
         <div class="likes">
@@ -64,18 +44,18 @@ export const firebaseLeerPublicacion = async () => {
 };
 
 export const firebaseDarLike = async (id) => {
-  const auth = getAuth();
+  // guardo el usuario actual autenticado en user
   const user=auth.currentUser
-  const db = getFirestore(FirebaseApp);
+  // inserto la sub coleccion likes con mi correo
   await addDoc(collection(doc(db, "Publicaciones", id), "likes"), {
     email: user.email,
   });
 };
 
 export const firebaseQuitarLike = async (id) => {
-  const auth = getAuth();
-  const db = getFirestore(FirebaseApp);
+  // obtengo los documentos de la sub coleccion likes de una publicacion
   const querySnapshot = await getDocs(collection(db, "Publicaciones", id, "likes"));
+  // busco mi email en la sub coleccion likes y lo elimino
   querySnapshot.forEach((doc) => {
     if (doc.data().email === auth.currentUser.email) {
       deleteDoc(doc.ref);
